Move variable_aliases into src
